backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# LLM and AI dependencies
transformers==4.36.0
torch==2.1.0
accelerate==0.25.0
sentencepiece==0.1.99
protobuf==4.25.1

# RAG and embeddings
chromadb==0.4.18
langchain==0.1.0
sentence-transformers==2.2.2

# Sentiment analysis (FinBERT)
transformers[torch]

# Document processing
pypdf==3.17.0
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1

# Earnings call transcripts (optional - uses mock data if not available)
# For real data: pip install git+https://github.com/EarningsCall/earningscall-python.git
# earningscall

# Database and caching
redis==5.0.1
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
black==23.12.0
